+++ CHANGELOG.md
@@ -22,4 +22,4 @@ from fuseki. Remove temp file processing for loading og graph -- load directly. 
     * vivofoundation.py -- Support code for the pump.  Will be renamed.
 * **2015-02-06** 0.37 Minor code improvements to address pylint and pyCharm suggestions
 * **2015-02-06** 0.38 WORK IN PROGRESS.  New vivopump library.  New calls.  Updating code.  0.38 is not operational.
-* **2015-02-06** 0.39 pyunit tests for each function in vivopump.  Tests pass.  More tests can/should be added.  sv_orgs operational.  get and update tested.  Code now refers to vivopump for helper functions and uses new vivopump function names.
+* **2015-02-06** 0.39 pyunit tests for each function in vivopump.  Tests pass.  More tests can/should be added.  sv_orgs operational.  get and update tested.  Code now refers to vivopump for helper functions and uses new vivopump function names. prepare_column_values split out from do_get.  Added label and type triples for intermediate objects